// src/api/generated/models/orderStatus.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Nest Shop API
 * API for my startup
 * OpenAPI spec version: 1.0.0
 */

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED',
} as const;